Routing working for concerts

+++ client/controllers/concert.js
@@ -5,17 +5,19 @@ const cat = require("../api/cat");
 /* GET concerts listing. */
 exports.concerts_list = async function (req, res, next) {
   await concerts.findAll().then((concerts) => {
-    res.render("concerts", { title: "List of concerts", concerts: concerts });
+    res.render("concerts", { title: "List of concerts", concerts: concerts, token: req.cookies.jwt });
   });
 };
 
 /* GET concerts add */
 exports.concerts_create_get = function (req, res, next) {
-  res.render("concerts_input", { title: "Add a new concert" });
+  res.render("concerts_input", { title: "Add a new concert", token: req.cookies.jwt });
 };
 
 /* POST concerts add */
 exports.concerts_create_post = async function (req, res, next) {
+  if(req.cookies.jwt === (undefined || null || "")) return res.redirect("/login");
+
   var newConcert = {
     title: req.body.title,
     city: req.body.city,
@@ -48,13 +50,16 @@ exports.concerts_detail = async function (req, res, next) {
     res.render("concert", {
       title: concert.title,
       concert,
-      catFact
+      catFact,
+      token: req.cookies.jwt
     });
   });
 };
 
 /* DELETE concerts */
 exports.concerts_delete = async function (req, res, next) {
+  if(req.cookies.jwt === (undefined || null || "")) return res.redirect("/login");
+
   await concerts.deleteById(req.params.uuid).then(() => {
     res.redirect("/concerts");
   });
@@ -62,16 +67,21 @@ exports.concerts_delete = async function (req, res, next) {
 
 /* GET concerts edit */
 exports.concerts_edit_get = async function (req, res, next) {
+  if(req.cookies.jwt === (undefined || null || "")) return res.redirect("/login");
+
   await concerts.findById(req.params.uuid).then((concert) => {
     res.render("concerts_input", {
       title: `Edit ${concert.title}`,
       concert: concert,
+      token: req.cookies.jwt
     });
   });
 };
 
 /* POST concerts edit */
 exports.concerts_edit_put = async function (req, res, next) {
+  if(req.cookies.jwt === (undefined || null || "")) return res.redirect("/login");
+
   var updatedConcert = {
     _id: req.params.uuid,
     title: req.body.title,
